fix/test: t.Fatal should be called from running the test or benchmark func
The linter complains: the goroutine calls T.Fatal, which must be called in the
same goroutine as the testSA2002(default).

testing.T.Fail will calls testing.common.FailNow(), this function will marks
the test failing and call runtime.GoExit(). Sure, it could let the calling
goroutine exit, but this is a misuse.

t.Fail should be used for marks the test failing and let the goroutine which
runs this test or the benchmark func exit. So, just use t.Error+return instead.